requirements.txt
Flask==2.0.1
geopy==2.2.0
google-cloud-storage==2.0.0
python-dotenv==0.19.1
gunicorn==20.1.0
